feature(gh-91): Split WAF code into its own file
